lms/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
python-decouple==3.8
drf-spectacular==0.27.2
django-filter==23.5
setuptools>=69.0.0